пр14 done

+++ controllers/cards.js
@@ -3,50 +3,48 @@ const {
   STATUS_OK,
   BAD_REQUEST,
   NOT_FOUND,
-  INTERNAL_SERVER_ERROR,
+  FORBIDDEN,
 } = require('../utils/consts');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => {
       if (cards.length > 0) res.status(STATUS_OK).send(cards);
       else res.status(NOT_FOUND).send({ message: 'Карточки не найдены' });
     })
-    .catch(() => {
-      res.stats(INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+    .catch((err) => {
+      next(err);
     });
 };
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body;
   return Card.create({ name, link, owner: req.user._id })
     .then(() => res.status(STATUS_OK).send(req.body))
     .catch((err) => {
-      if (err.name === 'ValidationError') {
-        res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные при создании карточки' });
-      } else res.status(INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  return Card.findByIdAndDelete(cardId)
-    .then((result) => {
-      console.log('result', result);
-
-      if (result !== null) {
-        res.status(STATUS_OK).send({ message: 'карточка удалена' });
+  return Card.findById(cardId)
+    .then((card) => {
+      if (card !== null) {
+        if (card.owner.toString() === req.user._id) {
+          Card.findByIdAndDelete(cardId)
+            .then(() => {
+              res.status(STATUS_OK).send({ message: 'карточка удалена' });
+            });
+        } else res.status(FORBIDDEN).send({ message: 'Недостаточно прав доступа' });
       } else res.status(NOT_FOUND).send({ message: 'Карточка с указанным _id не найдена' });
     })
     .catch((err) => {
-      console.log('err catch del', err);
-      if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'переданы некорректные данные' });
-      } else res.status(INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
 
-const addLike = (req, res) => {
+const addLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
@@ -58,13 +56,11 @@ const addLike = (req, res) => {
       } else res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки лайка' });
     })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'переданы некорректные данные' });
-      } else res.status(INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
 
-const deleteLike = (req, res) => {
+const deleteLike = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } }, // убрать _id из массива
@@ -76,9 +72,7 @@ const deleteLike = (req, res) => {
       } else res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные для снятия лайка' });
     })
     .catch((err) => {
-      if (err.name === 'CastError') {
-        res.status(BAD_REQUEST).send({ message: 'переданы некорректные данные' });
-      } else res.status(INTERNAL_SERVER_ERROR).send({ message: 'На сервере произошла ошибка' });
+      next(err);
     });
 };
